Crate wiggle_runtime

Structs

GuestBorrows

GuestPtr
A guest pointer into host memory.

Region
Represents a contiguous region in memory.

Enums

GuestError

Traits

GuestErrorType

GuestMemory
A trait which abstracts how to get at the region of host memory taht contains guest memory.

GuestType
A trait for types that are intended to be pointees in GuestPtr<T>.

GuestTypeTransparent
A trait for GuestTypes that have the same representation in guest memory as in Rust. These types can be used with the GuestPtr::as_raw method to view as a slice.

Pointee
Types that can be pointed to by GuestPtr<T>.